pyproject.toml
[project]
name = "cathayragbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.77.0",
    "gradio>=6.9.0",
    "llama-index>=0.14.15",
    "llama-index-embeddings-openai>=0.5.2",
    "llama-index-llms-openai>=0.6.26",
    "llama-index-readers-docling>=0.4.2",
    "python-dotenv>=1.2.2",
    "trulens-apps-llamaindex>=2.7.0",
    "trulens-eval>=2.7.0",
    "trulens-providers-openai>=2.7.0",
]
